fix: route openSymbol through shared module, handle in AgentManagerProvider too
The symbol click in the Agent Manager panel (tab view) was silently
dropped because messages from that webview go to AgentManagerProvider
not KiloProvider. Extract handleOpenSymbol into open-symbol.ts and
wire it into both providers.
